# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banking App OpenApi",
        "description": "Banking App API Explorer..",
        "contact": {
            "email": "gladwell_n@live.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Swagger OpenApi server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration.",
                "description": "Creates new users.",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Gladwell Ndlovu",
                                    "email": "gladwell_n@live.com",
                                    "password": "P@ssword01",
                                    "password_confirmation": "P@ssword01"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/email/verify/{id}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Email verification.",
                "description": "Marks user email as verified.",
                "operationId": "emailVerify",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/email/resend": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Resend email verification link.",
                "description": "Sends email verification link.",
                "operationId": "emailResend",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login.",
                "description": "Authenticates users.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com",
                                    "password": "P@ssword01"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/two-factor": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Google2FA login.",
                "description": "Validate Google2FA code and login..",
                "operationId": "twoFactorLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    },
                                    "code": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com",
                                    "code": 123456
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh authentication token..",
                "description": "Invalidates old tokens and return a new one.",
                "operationId": "tokenRefresh",
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User profile",
                "description": "Returns user profile.",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User profile",
                "description": "Invalidate authentication token.",
                "operationId": "profile",
                "responses": {
                    "201": {
                        "description": "Successful request"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/password/forgot": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Password reset link.",
                "description": "Sends password reset link.",
                "operationId": "forgotPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "gladwell_n@live.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful request"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "429": {
                        "description": "Too Many Requests"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/password/reset": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password",
                "description": "Saves new user password",
                "operationId": "resetPassword",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/hello": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Simple test route",
                "description": "Returns my name.",
                "operationId": "testRoute",
                "responses": {
                    "200": {
                        "description": "Successful request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste the token below:",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Test",
            "description": "API Endpoints for testing."
        },
        {
            "name": "Authentication",
            "description": "API Endpoints for user authentication."
        }
    ],
    "security": [
        []
    ]
}